Backport multiple fixes (#78)
* Prevent overwrites due to rebalance-stop race (#20233)

Rebalance-stop can race with ongoing rebalance operations. This change
prevents these operations from overwriting objects by checking the source
and destination pool indices are different.

* avoid source index to be same as destination index (#20238)

during rebalance stop, it can possibly happen that
Put() would race by overwriting the same object again.

This may very well if done "successfully" it can
potentially proceed to delete the object from the pool,
causing data loss.

This PR enhances #20233 to handle more scenarios such
as these.

* remove windows CI/CD for now (#20441)

windows has decided to be a community support
only and source compile-friendly.

* DeleteObjects: Send delete to all pools (#172) (#20821)

Currently, DeleteObjects() tries to find the object's pool before
sending a delete request. This only works well when an object has
multiple versions in different pools since looking for the pool does
not consider the version-id. When an S3 client wants to
remove a version-id that exists in pool 2, the delete request will be
directed to pool one because it has another version of the same object.

This commit will remove looking for pool logic and will send a delete
request to all pools in parallel. This should not cause any performance
regression in most of the cases since the object will unlikely exist
in only one pool, and the performance price will be similar to
getPoolIndex() in that case.

* do not list buckets without local quorum (#20852)

ListBuckets() would result in listing buckets
without quorum, this PR fixes the behavior.

* fix: site replication of bucket deletion sync (#352)

Bucket deletion timestamp was not being passed back
in GetBucketInfo, which is needed to decide on the bucket
creation/deletion

* heal: Avoid retrying healing the whole erasure set

* fix: proxy requests to honor global transport

* fix: proxy requests to honor global transport
Load the globalProxyEndpoint properly

also, currently, the proxy requests will fail silently for batch cancel
even if the proxy fails; instead,d properly send the corresponding error back
for such proxy failures if opted

* pass the transport to the GetProxyEnpoints function

---------

Co-authored-by: Praveen raj Mani <praveen@minio.io>

---------

Co-authored-by: Krishnan Parthasarathi <krisis@users.noreply.github.com>
Co-authored-by: Harshavardhana <harsha@minio.io>
Co-authored-by: Poorna <poornas@users.noreply.github.com>
Co-authored-by: Anis Eleuch <anis@min.io>
Co-authored-by: Praveen raj Mani <praveen@minio.io>