heal: Avoid recreating a bucket without quorum (#460)
HealBucket can recreate a bucket if it is called without
opt.Remove flag.

Change the behavior by making HealBucket decides if a
bucket is stale or not. A bucket is considered stale if
every pool thinks the bucket is stale as well and it will
be removed (without force flag)