add CommitXL API for safer PUTs on unversioned buckets (#670)
implements and provides multi-stage protection on xl.meta
updates that involve

- inline data
- with older data-dir and xl.meta.bkp

under very rare scenarios there are possibility of incorrect
state() on the system after a 499 from the client on a PUT
request.

Bonus: also taken the liberty of simplifying and moving
away from from legacy format of 'xl.json' and call it
unsupported and ask for a migration via `mc mirror`.

from miniohq/eos
commit 4f87f6010cd66c7fa8097c4daef0b2c38823cef7 (HEAD)
Author: Harshavardhana <harsha@minio.io>
Date:   Thu Apr 17 15:28:39 2025 -0700