s3: DeleteObject to succeed when no enough meta quorum AND one offline node (#854)
Removing a non-versioned object when this latter does not have enough
quorum and there is one or more offline nodes can return 503, which is
unexpected.

The reason is that the type of InsufficientReadQuorum should be
calculated as RQInconsistentMeta instead of RQInsufficientOnlineDrives
in this specific case, so to allow DeleteObject() to continue removing
this object.